+++ src/user/user.service.ts
@@ -4,9 +4,9 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { User } from '../inteface/interface';
-import { CreateUserDto } from './dto/create-user-dto';
+import { CreateUserDto } from './dto/create-user.dto';
 import { v4 as uuidv4 } from 'uuid';
-import { UpdateUserDto } from './dto/update-user-dto';
+import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UserService {
